+++ src/Main.js
@@ -19,8 +19,12 @@ define(function (require, exports) {
         Setup             = require("src/utils/Setup"),
         Utils             = require("src/Utils");
 
-    var $icon                   = $("<a id='git-toolbar-icon' href='#'></a>").attr("title", Strings.LOADING)
-                                    .addClass("loading").appendTo($("#main-toolbar .buttons"));
+    var CMD_ADD_TO_IGNORE      = "git.addToIgnore",
+        CMD_REMOVE_FROM_IGNORE = "git.removeFromIgnore",
+        $icon                  = $("<a id='git-toolbar-icon' href='#'></a>")
+                                    .attr("title", Strings.LOADING)
+                                    .addClass("loading")
+                                    .appendTo($("#main-toolbar .buttons"));
 
     // This only launches when Git is available
     function initUi() {
@@ -105,30 +109,53 @@ define(function (require, exports) {
                 ErrorHandler.showError(expected, Strings.CHECK_GIT_SETTINGS);
             });
 
-            // add command to project menu
-            var projectCmenu = Menus.getContextMenu(Menus.ContextMenuIds.PROJECT_MENU);
-            var workingCmenu = Menus.getContextMenu(Menus.ContextMenuIds.WORKING_SET_MENU);
+            // register commands for project tree / working files
+            CommandManager.register(Strings.ADD_TO_GITIGNORE, CMD_ADD_TO_IGNORE, addItemToGitingore);
+            CommandManager.register(Strings.REMOVE_FROM_GITIGNORE, CMD_REMOVE_FROM_IGNORE, removeItemFromGitingore);
+
+            // create context menu for git panel
             var panelCmenu = Menus.registerContextMenu("git-panel-context-menu");
-            projectCmenu.addMenuDivider();
-            workingCmenu.addMenuDivider();
-
-            var cmdName = "git.addToIgnore";
-            CommandManager.register(Strings.ADD_TO_GITIGNORE, cmdName, addItemToGitingore);
-            projectCmenu.addMenuItem(cmdName);
-            workingCmenu.addMenuItem(cmdName);
-            CommandManager.register(Strings.ADD_TO_GITIGNORE, cmdName + "2", addItemToGitingoreFromPanel);
-            panelCmenu.addMenuItem(cmdName + "2");
-
-            cmdName = "git.removeFromIgnore";
-            CommandManager.register(Strings.REMOVE_FROM_GITIGNORE, cmdName, removeItemFromGitingore);
-            projectCmenu.addMenuItem(cmdName);
-            workingCmenu.addMenuItem(cmdName);
-            CommandManager.register(Strings.REMOVE_FROM_GITIGNORE, cmdName + "2", removeItemFromGitingoreFromPanel);
-            panelCmenu.addMenuItem(cmdName + "2");
+            CommandManager.register(Strings.ADD_TO_GITIGNORE, CMD_ADD_TO_IGNORE + "2", addItemToGitingoreFromPanel);
+            CommandManager.register(Strings.REMOVE_FROM_GITIGNORE, CMD_REMOVE_FROM_IGNORE + "2", removeItemFromGitingoreFromPanel);
+            panelCmenu.addMenuItem(CMD_ADD_TO_IGNORE + "2");
+            panelCmenu.addMenuItem(CMD_REMOVE_FROM_IGNORE + "2");
         });
     }
 
+    var _toggleMenuEntriesState = false,
+        _divider1 = null,
+        _divider2 = null;
+    function toggleMenuEntries(bool) {
+        if (bool === _toggleMenuEntriesState) {
+            return;
+        }
+        var projectCmenu = Menus.getContextMenu(Menus.ContextMenuIds.PROJECT_MENU);
+        var workingCmenu = Menus.getContextMenu(Menus.ContextMenuIds.WORKING_SET_MENU);
+        if (bool) {
+            _divider1 = projectCmenu.addMenuDivider();
+            _divider2 = workingCmenu.addMenuDivider();
+            projectCmenu.addMenuItem(CMD_ADD_TO_IGNORE);
+            workingCmenu.addMenuItem(CMD_ADD_TO_IGNORE);
+            projectCmenu.addMenuItem(CMD_REMOVE_FROM_IGNORE);
+            workingCmenu.addMenuItem(CMD_REMOVE_FROM_IGNORE);
+        } else {
+            projectCmenu.removeMenuDivider(_divider1.id);
+            workingCmenu.removeMenuDivider(_divider2.id);
+            projectCmenu.removeMenuItem(CMD_ADD_TO_IGNORE);
+            workingCmenu.removeMenuItem(CMD_ADD_TO_IGNORE);
+            projectCmenu.removeMenuItem(CMD_REMOVE_FROM_IGNORE);
+            workingCmenu.removeMenuItem(CMD_REMOVE_FROM_IGNORE);
+        }
+        _toggleMenuEntriesState = bool;
+    }
+
     // Event handlers
+    EventEmitter.on(Events.GIT_ENABLED, function () {
+        toggleMenuEntries(true);
+    });
+    EventEmitter.on(Events.GIT_DISABLED, function () {
+        toggleMenuEntries(false);
+    });
     // TODO: investigate this event
     EventEmitter.on(Events.HANDLE_PROJECT_REFRESH, function () {
         $(ProjectManager).triggerHandler("projectRefresh");